Moved SettingsItem to its own file inside the Settings folder.
